+++ bin/find.js
@@ -1,7 +1,12 @@
 #!/usr/bin/env node
 
 const arg = require('arg')
-const { getSpecs, collectResults, findChangedFiles } = require('../src')
+const {
+  getSpecs,
+  collectResults,
+  findChangedFiles,
+  getTests,
+} = require('../src')
 const { pickTaggedTestsFrom, leavePendingTestsOnly } = require('../src/tagged')
 const { addCounts } = require('../src/count')
 const { stringAllInfo } = require('../src/print')
@@ -60,94 +65,104 @@ debug('arguments %o', args)
 
 const specs = getSpecs()
 if (args['--names'] || args['--tags']) {
-  if (!specs.length) {
-    console.log('no specs found')
-  } else {
-    console.log('')
-    // counts the number of tests for each tag across all specs
-    const tagTestCounts = {}
-    const jsonResults = {}
+  // counts the number of tests for each tag across all specs
+  const { jsonResults, tagTestCounts } = getTests(specs, {
+    tags: args['--tags'],
+    tagged: args['--tagged'],
+    skipped: args['--skipped'],
+  })
+  // if (!specs.length) {
+  //   console.log('no specs found')
+  // } else {
+  //   console.log('')
+  // counts the number of tests for each tag across all specs
+  // const tagTestCounts = {}
+  // const jsonResults = {}
 
-    specs.forEach((filename) => {
-      jsonResults[filename] = {
-        counts: {
-          tests: 0,
-          pending: 0,
-        },
-        tests: [],
-      }
-      const source = fs.readFileSync(filename, 'utf8')
-      const result = getTestNames(source, true)
-      // enable if need to debug the parsed test
-      // console.dir(result.structure, { depth: null })
-      collectResults(result.structure, jsonResults[filename].tests)
+  // specs.forEach((filename) => {
+  //   jsonResults[filename] = {
+  //     counts: {
+  //       tests: 0,
+  //       pending: 0,
+  //     },
+  //     tests: [],
+  //   }
+  //   const source = fs.readFileSync(filename, 'utf8')
+  //   const result = getTestNames(source, true)
+  //   // enable if need to debug the parsed test
+  //   // console.dir(result.structure, { depth: null })
+  //   collectResults(result.structure, jsonResults[filename].tests)
 
-      if (args['--tags']) {
-        const specTagCounts = countTags(result.structure)
-        Object.keys(specTagCounts).forEach((tag) => {
-          if (!(tag in tagTestCounts)) {
-            tagTestCounts[tag] = specTagCounts[tag]
-          } else {
-            tagTestCounts[tag] += specTagCounts[tag]
-          }
-        })
-      }
-    })
+  //   if (args['--tags']) {
+  //     const specTagCounts = countTags(result.structure)
+  //     Object.keys(specTagCounts).forEach((tag) => {
+  //       if (!(tag in tagTestCounts)) {
+  //         tagTestCounts[tag] = specTagCounts[tag]
+  //       } else {
+  //         tagTestCounts[tag] += specTagCounts[tag]
+  //       }
+  //     })
+  //   }
+  // })
 
-    addCounts(jsonResults)
+  // addCounts(jsonResults)
 
-    if (args['--names']) {
-      if (args['--tagged']) {
-        // filter all collected tests to those that have the given tag(s)
-        const splitTags = args['--tagged']
-          .split(',')
-          .map((s) => s.trim())
-          .filter(Boolean)
-        debug('filtering all tests by tag "%o"', splitTags)
-        pickTaggedTestsFrom(jsonResults, splitTags)
-        // recompute the number of tests
-        addCounts(jsonResults)
-      } else if (args['--skipped']) {
-        debug('leaving only skipped (pending) tests')
-        leavePendingTestsOnly(jsonResults)
-        // recompute the number of tests
-        addCounts(jsonResults)
-      }
+  if (args['--names']) {
+    // const jsonResults = getTests({
+    //   tagged: args['--tagged'],
+    //   skipped: args['--skipped'],
+    // })
+    // if (args['--tagged']) {
+    //   // filter all collected tests to those that have the given tag(s)
+    //   const splitTags = args['--tagged']
+    //     .split(',')
+    //     .map((s) => s.trim())
+    //     .filter(Boolean)
+    //   debug('filtering all tests by tag "%o"', splitTags)
+    //   pickTaggedTestsFrom(jsonResults, splitTags)
+    //   // recompute the number of tests
+    //   addCounts(jsonResults)
+    // } else if (args['--skipped']) {
+    //   debug('leaving only skipped (pending) tests')
+    //   leavePendingTestsOnly(jsonResults)
+    //   // recompute the number of tests
+    //   addCounts(jsonResults)
+    // }
 
-      if (args['--count']) {
-        let n = 0
-        Object.keys(jsonResults).forEach((filename) => {
-          const skippedCount = jsonResults[filename].counts.pending
-          n += skippedCount
-        })
-        console.log(n)
-      } else {
-        if (args['--json']) {
-          console.log(JSON.stringify(jsonResults, null, 2))
-        } else {
-          const str = stringAllInfo(jsonResults)
-          console.log(str)
-        }
-        console.log('')
-      }
-    }
-
-    if (args['--tags']) {
-      const tagEntries = Object.entries(tagTestCounts)
-      const sortedTagEntries = tagEntries.sort((a, b) => {
-        // every entry is [tag, count], so compare the tags
-        return a[0].localeCompare(b[0])
+    if (args['--count']) {
+      let n = 0
+      Object.keys(jsonResults).forEach((filename) => {
+        const skippedCount = jsonResults[filename].counts.pending
+        n += skippedCount
       })
+      console.log(n)
+    } else {
       if (args['--json']) {
-        // assemble a json object with the tag counts
-        const tagResults = Object.fromEntries(sortedTagEntries)
-        console.log(JSON.stringify(tagResults, null, 2))
+        console.log(JSON.stringify(jsonResults, null, 2))
       } else {
-        const table = consoleTable.getTable(['Tag', 'Tests'], sortedTagEntries)
-        console.log(table)
+        const str = stringAllInfo(jsonResults)
+        console.log(str)
       }
+      console.log('')
     }
   }
+
+  if (args['--tags']) {
+    const tagEntries = Object.entries(tagTestCounts)
+    const sortedTagEntries = tagEntries.sort((a, b) => {
+      // every entry is [tag, count], so compare the tags
+      return a[0].localeCompare(b[0])
+    })
+    if (args['--json']) {
+      // assemble a json object with the tag counts
+      const tagResults = Object.fromEntries(sortedTagEntries)
+      console.log(JSON.stringify(tagResults, null, 2))
+    } else {
+      const table = consoleTable.getTable(['Tag', 'Tests'], sortedTagEntries)
+      console.log(table)
+    }
+  }
+  // }
 } else if (args['--branch']) {
   debug('determining specs changed against branch %s', args['--branch'])
   let changedFiles = findChangedFiles(args['--branch'], args['--parent'])
