+++ package.json
@@ -14,7 +14,7 @@
     "test": "ava",
     "cy:run": "DEBUG=cypress:cli,cypress:server:specs cypress run",
     "demo": "DEBUG=find-cypress-specs node ./bin/find",
-    "demo-names": "node ./bin/find --names",
+    "demo-names": "DEBUG=find-cypress-specs node ./bin/find --names",
     "demo-skipped-tests": "node ./bin/find --names --skipped",
     "demo-count-skipped-tests": "node ./bin/find --names --skipped --count",
     "demo-tags": "node ./bin/find --tags",
